Move ToggleSwitch to @rancher/components
