+++ pkg/rancher-components/src/components/Form/ToggleSwitch/ToggleSwitch.vue
