+++ pkg/rancher-components/src/components/Form/ToggleSwitch/index.ts
@@ -1,0 +1,1 @@
+export { default as ToggleSwitch } from './ToggleSwitch.vue';
