+++ pkg/rancher-components/src/components/Form/index.ts
@@ -2,3 +2,4 @@ export * from './Checkbox';
 export * from './LabeledInput';
 export * from './Radio';
 export * from './TextArea';
+export * from './ToggleSwitch';
